+++ lib/src/card/card_model.dart
@@ -1,0 +1,10 @@
+/*
+ * Copyright (c) TIKI Inc.
+ * MIT license. See LICENSE file in root directory.
+ */
+
+class CardModel {
+  double balance;
+
+  CardModel({this.balance = 0});
+}