+++ lib/src/cash_out/cash_out_view_widget_banner.dart
@@ -1,0 +1,35 @@
+/*
+ * Copyright (c) TIKI Inc.
+ * MIT license. See LICENSE file in root directory.
+ */
+import 'package:flutter/material.dart';
+import 'package:tiki_style/tiki_style.dart';
+
+class CashOutViewWidgetBanner extends StatelessWidget {
+  static const String _text = "Coming soon";
+
+  const CashOutViewWidgetBanner({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Stack(
+      children: [
+        SizedBox(
+          child: ImgProvider.bannerCyan,
+          height: SizeProvider.instance.size(30),
+        ),
+        Container(
+            margin: EdgeInsets.symmetric(
+                vertical: SizeProvider.instance.size(4),
+                horizontal: SizeProvider.instance.size(12)),
+            child: Text(_text,
+                style: TextStyle(
+                    fontWeight: FontWeight.bold,
+                    fontFamily: TextProvider.familyNunitoSans,
+                    package: TextProvider.package,
+                    fontSize: SizeProvider.instance.text(16),
+                    color: ColorProvider.tikiBlue)))
+      ],
+    );
+  }
+}