+++ lib/statement/model.dart
@@ -1,5 +1,0 @@
-import '../transaction/model/money_model_transaction.dart';
-
-class StatementModel{
-  List<TransactionModel> transactions = [];
-}