+++ lib/tiki_money.dart
@@ -1,26 +1,30 @@
+/*
+ * Copyright (c) TIKI Inc.
+ * MIT license. See LICENSE file in root directory.
+ */
+
 import 'package:flutter/material.dart';
-import 'package:httpp/httpp.dart';
+import 'package:tiki_localgraph/tiki_localgraph.dart';
+import 'package:tiki_wallet/tiki_wallet.dart';
 
-export 'transaction/model/money_model_transaction.dart';
+import 'src/card/card_service.dart';
+import 'src/nft_history/nft_history_service.dart';
+import 'src/screen/screen_service.dart';
 
-import 'container/service.dart';
-import 'transaction/model/money_model_transaction.dart' as transaction;
+export 'src/nft_detail/nft_detail_model.dart';
+export 'src/nft_detail/nft_detail_model_type.dart';
 
 class TikiMoney {
-  late final ContainerService _service;
+  late final ScreenService _screenService;
 
-  TikiMoney({Httpp? httpp, required int referalCount,
-    List<transaction.TransactionModel> transactions = const []})
-      : _service = ContainerService(
-            httpp: httpp ?? Httpp(),
-            referalCount: referalCount) {
-    pushTransactions(transactions);
+  TikiMoney(
+      {required TikiLocalGraph localGraph,
+      required TikiChainService chainService,
+      int referralCount = 0}) {
+    _screenService = ScreenService(
+        card: CardService(referralCount: referralCount),
+        history: NFTHistoryService(localGraph, chainService));
   }
 
-  Widget screen({bool example = false}) =>
-      _service.presenter.home(example: example);
-
-  void pushTransactions(List<transaction.TransactionModel> transactions,
-      {bool overwrite = false}) =>
-        _service.statement.pushTransactions(transactions, overwrite: overwrite);
+  Widget get screen => _screenService.presenter.screen;
 }